Added query handler inside handler.
