+++ src/components/core/handlers/handler.ts
@@ -5,7 +5,8 @@ import {
   NonceCommand,
   GetFeesCommand,
   Command,
-  EncryptCommand
+  EncryptCommand,
+  QueryCommand
 } from '../../../utils/constants.js'
 import {
   DB_CONSOLE_LOGGER,
@@ -239,6 +240,37 @@ export class EncryptHandler extends Handler {
       const encryptedData = await encrypt(blobData, this.getTask().encryptionType)
       return {
         stream: Readable.from(encryptedData.toString('hex')),
+        status: { httpStatus: 200 }
+      }
+    } catch (error) {
+      return {
+        stream: null,
+        status: { httpStatus: 500, error: 'Unknown error: ' + error.message }
+      }
+    }
+  }
+}
+
+export class QueryHandler extends Handler {
+  public constructor(task: any, database: Database) {
+    super(task, null, database)
+    if (!this.isQueryCommand(task)) {
+      throw new Error(`Task has not QueryCommand type. It has ${typeof task}`)
+    }
+  }
+
+  isQueryCommand(obj: any): obj is QueryCommand {
+    return typeof obj === 'object' && obj !== null && 'command' in obj && 'query' in obj
+  }
+
+  async handle(): Promise<P2PCommandResponse> {
+    try {
+      let result = await this.getDatabase().ddo.search(this.getTask().query)
+      if (!result) {
+        result = []
+      }
+      return {
+        stream: Readable.from(JSON.stringify(result)),
         status: { httpStatus: 200 }
       }
     } catch (error) {
